+++ src/js/directives/scrollspy.js
@@ -18,9 +18,12 @@ export default {
   inserted(el, binding, vnode) {
     let onActivate = (item) => {
       vnode.context.$root.$emit('scrollspy-activate', item);
+      if (binding.value && typeof binding.value.onActivate === 'function') {
+        binding.value.onActivate(item);
+      }
     };
 
-    let options = Object.assign({}, { onActivate }, binding.value);
+    let options = Object.assign({}, binding.value, { onActivate });
     let scrollspy = new Scrollspy(el, options);
 
     vnode.context.$root.$on('scrollspy-refresh', () => scrollspy.refresh());